model Terrace change longitude and latitude into float

+++ db/migrate/20200217143415_create_terraces.rb
@@ -4,8 +4,8 @@ class CreateTerraces < ActiveRecord::Migration[5.2]
       t.string :name
       t.text :description
       t.string :address
-      t.integer :longitude
-      t.integer :lattitude
+      t.float :longitude
+      t.float :lattitude
       t.string :status
       t.references :user, foreign_key: true
